renamed Sidebar file; review page rendered with book info in-progress
