+++ client/components/AddReviewForm.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { connect } from 'react-redux';
 
 class AddReviewForm extends React.Component {
   constructor () {
@@ -17,9 +18,18 @@ class AddReviewForm extends React.Component {
   }
 
   render () {
+    if (this.props.books.length === 0){
+      return <div>Loading...</div>
+    }
+
+    let currentBook = this.props.books.filter((book) => {
+      return (book.id === +this.props.match.params.bookId)
+    })
+    console.log(currentBook[0])
+
     return (
       <div id="review-form-container">
-        <h3>Tell your story</h3>
+        <h3>{`Tell us about ${currentBook[0].title} by ${currentBook[0].author.firstName} ${currentBook[0].author.lastName}`}</h3>
         <form id="review-form" >
           <label htmlFor="rating"> Rating </label>
             <select>
@@ -40,4 +50,10 @@ class AddReviewForm extends React.Component {
   }
 }
 
-export default AddReviewForm
+const mapState = state => {
+  return {
+    books: state.books
+  }
+}
+
+export default connect(mapState, null)(AddReviewForm)
