+++ client/components/Sidebar.js
@@ -1,18 +1,18 @@
 import React from "react";
-import { connect } from "react-redux";
+import { Link } from "react-router-dom";
 
-export default class Sidebar extends React.Component {
-  render () {
-    return (
-    <div>
+const Sidebar = () => (
+  <div id="sidebar-container">
+    <div id="sidebar-links">
+    <Link to="/"><div className="link-name top-link">Books</div></Link>
+     <Link to="/authors"><div className="link-name">Authors</div></Link>
+
+      <Link to="/genres"><div className="link-name">Genres</div></Link>
+
+      <Link to="/cart"><div className="link-name">Cart</div></Link>
 
     </div>
-    )
-  }
-}
+  </div>
+)
 
-// const mapState = state => ({
-//   books: state.books
-// });
-
-// export default connect(mapState, null)(Sidebar);
+export default Sidebar
