+++ client/components/index.js
@@ -7,6 +7,6 @@ export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
 export {Login, Signup} from './auth-form'
 export {default as AllBooks} from './AllBooks'
-export {default as Sidebar} from './sidebar'
+export {default as Sidebar} from './Sidebar'
 export {default as AllAuthors} from './AllAuthors'
 export {default as AddReviewForm} from './AddReviewForm'
